requirements.txt
requests[socks]>=2.31.0  # Ajustado para incluir el soporte SOCKS5
telebot>=0.0.5
pyrogram>=2.0.106
telethon>=1.33.1
aiohttp==3.10.5
pymongo[srv]==4.7.0
random-address>=1.1.1
colored>=2.2.4
colorama>=0.4.6
lxml>=4.9.3
cloudscraper==1.2.71
bs4>=0.0.1
telegram>=0.0.1
telegrapi>=1.0.0
psutil>=5.9.7
TgCrypto
python-dotenv==1.0.0
names
flask==2.0.3
pymongo==4.7.0
beautifulsoup4==4.12.3
capsolver==1.0.7
faker
autopep8==2.3.1
fake-useragent
retry  # Ya está incluido
PySocks  # Ya está incluido, necesario para requests[socks]
playwright
pyrate-limiter==3.6.1
sentry-sdk
pytz
httpx>=0.27.0
motor>=3.3.1
aiofiles
httpx_socks
aiohttp_socks==0.8.4
urllib3>=1.26.0
cloudscraper
curl-cffi
capsolver-python
cachetools==5.5.0
asyncio-throttle==1.0.2
tenacity==9.0.0
redis==5.1.1
aioredis
pycryptodome==3.20.0
motor>=3.5.0
prometheus_client>=0.20.0